requirements.txt
flask==0.12.2
flask-cors==3.0.3
flask-restful==0.3.6
flask-session==0.3.1
flask-admin==1.5.0
flask-login==0.4.0
Flask-WTF==0.14.2
PyJWT==1.6.0
PyMySQL==0.8.0
peewee==2.10.2
wtf-peewee==0.2.6
redis==2.10.5
passlib==1.7.1
uwsgi==2.0.15
gevent==1.2.2
scrypt==0.8.0
fabric==1.14.0
fabric3==1.14.post1